fixes, cr feedback, organizing double_representation module
