+++ lib/ide/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2018"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-ast                      = { version = "0.1.0"   , path = "ast/impl"        }
+ast                      = { version = "0.1.0"  , path = "ast/impl"         }
 basegl                   = { version = "0.1.0"  , path = "../core"          }
 basegl-core-msdf-sys     = { version = "0.1.0"  , path = "../core/msdf-sys" }
 basegl-system-web        = { version = "0.1.0"  , path = "../system/web"    }
